Add r_sys_now_mono() and use in r2r ##util (#17276)
